backend/requirements.txt

fastapi

uvicorn

psycopg2-binary

groq

pydantic

